@@ -1,8 +1,8 @@
 package com.xwray.groupie.example.databinding;
 
 import android.graphics.drawable.Animatable;
-import android.support.annotation.NonNull;
-import android.support.annotation.StringRes;
+import androidx.annotation.NonNull;
+import androidx.annotation.StringRes;
 import android.view.View;
 
 import com.xwray.groupie.ExpandableGroup;
