@@ -1,7 +1,7 @@
 package com.xwray.groupie.example.databinding;
 
-import android.support.annotation.ColorInt;
-import android.support.annotation.LayoutRes;
+import androidx.annotation.ColorInt;
+import androidx.annotation.LayoutRes;
 
 public class HeaderItemDecoration extends com.xwray.groupie.example.core.decoration.HeaderItemDecoration {
     public HeaderItemDecoration(@ColorInt int background, int sidePaddingPixels) {
