@@ -1,7 +1,7 @@
 package com.xwray.groupie.example.databinding;
 
-import android.support.annotation.ColorInt;
-import android.support.annotation.Dimension;
+import androidx.annotation.ColorInt;
+import androidx.annotation.Dimension;
 
 public class InsetItemDecoration extends com.xwray.groupie.example.core.decoration.InsetItemDecoration {
     public InsetItemDecoration(@ColorInt int backgroundColor, @Dimension int padding) {
