@@ -1,6 +1,6 @@
 package com.xwray.groupie.example
 
-import android.support.annotation.ColorInt
+import androidx.annotation.ColorInt
 
 class HeaderItemDecoration(@ColorInt background: Int, sidePaddingPixels: Int)
     : com.xwray.groupie.example.core.decoration.HeaderItemDecoration(
