@@ -1,6 +1,6 @@
 package com.xwray.groupie.example.item
 
-import android.support.annotation.ColorInt
+import androidx.annotation.ColorInt
 import com.xwray.groupie.example.INSET
 import com.xwray.groupie.example.INSET_TYPE_KEY
 
