@@ -1,6 +1,6 @@
 package com.xwray.groupie.example.item
 
-import android.support.annotation.ColorRes
+import androidx.annotation.ColorRes
 import com.xwray.groupie.example.INSET_TYPE_KEY
 
 class FullBleedCardItem(@ColorRes colorRes: Int) : CardItem(colorRes) {
