@@ -1,7 +1,7 @@
 package com.xwray.groupie.example.item
 
 import android.graphics.drawable.Animatable
-import android.support.annotation.ColorInt
+import androidx.annotation.ColorInt
 import com.xwray.groupie.example.INSET
 import com.xwray.groupie.example.INSET_TYPE_KEY
 import com.xwray.groupie.example.R
