@@ -1,6 +1,6 @@
 package com.xwray.groupie.example.item
 
-import android.support.annotation.ColorInt
+import androidx.annotation.ColorInt
 
 open class SmallCardItem(@ColorInt private val colorRes: Int, text: CharSequence? = "") : CardItem(colorRes, text) {
 
