@@ -1,6 +1,6 @@
 package com.xwray.groupie.example.item
 
-import android.support.annotation.ColorRes
+import androidx.annotation.ColorRes
 
 data class UpdatableItem(@ColorRes private val colorRes: Int, private val index: Int) : SmallCardItem(colorRes, index.toString()) {
 
