@@ -1,7 +1,7 @@
 package com.xwray.groupie.databinding;
 
-import android.databinding.ViewDataBinding;
-import android.support.annotation.NonNull;
+import androidx.databinding.ViewDataBinding;
+import androidx.annotation.NonNull;
 
 public class ViewHolder<T extends ViewDataBinding> extends com.xwray.groupie.ViewHolder {
     public final T binding;
