@@ -1,8 +1,8 @@
 package com.xwray.groupie;
 
-import android.support.annotation.NonNull;
-import android.support.v7.widget.RecyclerView;
-import android.support.v7.widget.helper.ItemTouchHelper;
+import androidx.annotation.NonNull;
+import androidx.recyclerview.widget.RecyclerView;
+import androidx.recyclerview.widget.ItemTouchHelper;
 
 public abstract class TouchCallback extends ItemTouchHelper.SimpleCallback {
 
